chore: move EventGenerator to debug package
