+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/debug/EventGenerator.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-package com.wire.kalium.logic.sync
+package com.wire.kalium.logic.feature.debug
 
 import com.benasher44.uuid.uuid4
 import com.wire.kalium.cryptography.CryptoClientId
